requirements.txt
PyMuPDF==1.26.6
Pillow==11.0.0
tkinterdnd2==0.4.3
ghostscript==0.7
